[qca-hyfi-bridge] Add external API for ECM
hyfi_ecm_new_connection: Will be called upon a creation of a new
connection in ECM. Function receives the triggering skb and serial
number, and returns Hy-Fi hash which will be stored for future
references.

hyfi_ecm_update_stats: Will be called when ECM wants to updated
the stats on a specific connection. Function receives the serial
number and hash, as well as updated stats. This will keep the
Hy-Fi algorithms happy on accelerated flows.

Change-Id: Ia6df036938633b6e598c6dcf1fde3169a101d59b
Signed-off-by: Hai Shalom <hshalom@codeaurora.org>