[hyfi-netfilter] Avoid trivial hash collisions
In an attempt to make the source and destination ports less correlated
in the final hash, combine the destination port with the destination IP
address and the source port with the source IP address. This at least
helps when the port numbers are sequentially increasing (which
previously could lead to a collision rather easily).

Change-Id: If957e7c80a111ea5d80612101da84f041318d74b
Signed-off-by: Bharat Bhushan <bchakrav@codeaurora.org>