[hyfi-netfilter] Keep both forward and reverse hashes for ECM
NSS tracks a bi-directional connection whereas Hy-Fi tracks a
unidirectional flow (identified by a hash and a destination MAC
address). To facilitate mapping from the NSS connection to the
appropriate H-Active entries, compute two hash values at the time a flow
is accelerated. The first directly uses the fields in the skb to compute
the hash. The second swaps the source and destination addresses and
ports to arrive at the hash for the reverse direction of the connection.

Computing this reverse direction hash is done through a new hash
function. Due to potential performance impacts, this is a duplicate of
the original hash function with some of the logic removed that does not
apply for the hash computation itself.

Note that these changes now requires the SA, DA, and hash as input
parameters to hyfi_ecm_update_stats since these are different depending
on whether the flow is in the forward or reverse direction.

Also reset the byte and packet count if NSS serial value is changed.

Change-Id: I16cee6bac5fab2286ec8269b6bd7325b5a74aba9
Signed-off-by: Bharat Bhushan <bchakrav@codeaurora.org>