[hyfi-netfilter] Allow FDB updates with single non-relaying port
When there is only one non-relaying port, there should be no chance of a
loop causing a MAC address to be learned on the wrong interface.
Therefore, it is unnecessary to enforce the rule that says a MAC address
learned on a relaying interface cannot be re-learned on a non-relaying
interface.

To implement this, modify the bridge to keep track of the number
non-relaying interfaces. Only apply the FDB update check when this count
is greater than 1.

Change-Id: Id74d38dc46a7fedb3596cc5bc6c8f720fd6ee9bc
Signed-off-by: Bharat Bhushan Chakravarty <bchakrav@codeaurora.org>