qca-hyfi-bridge: Update SON/EasyMesh Version
1. Major version is incremented to 10 to ensure enough version
gap is available for 11.x cadence, assuming both major and
minor changes are expected in 11.x

2. Minor version is initialized to zero.

3. Intra-node compatibility (between SON application and UMAC driver)
is invalid. Because 11.x is 4.4 kernel and 12.x is 5.4 kernel.
So SON/EasyMesh package from 11.x cannot work in 12.x and vice versa

4. No need to increment compatibility number,
because 4.4 package cannot be installed in system running 5.4 kernel

5. Inter-node compatibility is maintained.
For example, Controller (11.5 – 1.0.0.0)
and Agent (12.0 – 10.0.0.0) are interoperable

Change-Id: Ie91c3c1d31df66cf138ebd2c17d830a6d3e187b0
Signed-off-by: Gnanaprakash Elumalai <quic_gelumala@quicinc.com>