qca-hyfi-bridge: MAP R5 Changes
For multiple HOP DPP we are seeing a delay in processing EAPOL
packets. The issue is due to hyfi-bridge taking lot of time to
process notification packets. On WFA setup we see lot of notifications
triggered due to hyManager.
Current Implementation and Changes made:
Current:
We install hooks in hyfi bridge to add the ifindex value into 1905
header . The adding of ifindex uses put_unaligned() API to copy the
ifindex into 1905 header. This process is consuming lot CPU cycles and
delaying packet processing . We see delay of upto 8 sec in 3 hop daisy.
New Soln:
We need not add ifindex in all packets . ezmesh just checks ifindex
on topology , vendor specific and multicast packets. So adding ifindex
check only on above mentioned packet types
Dual ALid mode changes were made to process packets for local forwarding.
Have added check to do this if AP is operating only in controller mode.
We do not need this on Agent

Change-Id: I8c4c330e0c2ce5a78d64927ea39dfb3f5495d6c9
Signed-off-by: Samanth Manjunath <quic_samamanj@quicinc.com>